Layout restructure: Move all the already existing GraphicsItems into viewer folder.
Before they were in common because there use was not clear.
