+++ src/ui/event/event_page_widget.h
@@ -24,7 +24,7 @@
 #include <QGraphicsScene>
 #include <lcf/rpg/eventcommand.h>
 #include <lcf/rpg/database.h>
-#include "ui/common/charset_item.h"
+#include "ui/viewer/charset_graphics_item.h"
 #include "ui/other/splash_dialog.h"
 #include <QProgressBar>
 #include <QLabel>
