+++ src/ui/picker/picker_child_widget.cpp
@@ -22,7 +22,7 @@ PickerChildWidget::PickerChildWidget(QWidget* parent) : QWidget(parent) {
 	// no-op
 }
 
-void PickerChildWidget::setView(ViewBase* view) {
+void PickerChildWidget::setView(RpgGraphicsView* view) {
 	assert(view);
 	m_view = view;
 }
