+++ src/ui/viewer/battle_animation_graphics_item.h
