+++ src/ui/viewer/charset_graphics_item.h
