+++ src/ui/viewer/stat_curve_graphics_item.h
